beast_requirements.txt
absl-py==0.15.0
affine==2.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
atomicwrites==1.1.5
attrs==21.4.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
black==22.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.22.2
colorama==0.4.3
contextlib2==21.6.0
cryptography==2.8
cycler==0.11.0
Cython==0.29.27
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
editdistance==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
fasteners==0.14.1
filterpy==1.4.5
Fiona==1.8.21
Flask==1.1.1
flatbuffers==1.12
fonttools==4.29.0
future==0.18.2
gast==0.3.3
geopandas==0.10.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==2.10.0
html5lib==1.0.1
httplib2==0.14.0
idna==3.3
imageio==2.14.1
importlib-metadata==4.10.1
importlib-resources==5.4.0
imutils==0.5.4
iniconfig==1.1.1
ipykernel==6.9.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.2
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
llvmlite==0.38.0
lmdb==1.3.0
lockfile==0.12.2
lxml==4.7.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nose==1.3.7
notebook==6.4.8
numba==0.55.1
numpy==1.19.0
oauthlib==3.2.0
object-detection==0.0.3
olefile==0.46
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.9
opt-einsum==3.3.0
OWSLib==0.19.1
p5py==1.0.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
patchify==0.2.3
path==16.3.0
pathspec==0.9.0
pbr==5.9.0
pdfkit==0.6.1
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.5.0
plotly==4.4.1
pluggy==1.0.0
progressbar2==4.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
protobuf==3.19.4
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.3.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyproj==2.5.0
PyQt5==5.14.1
PyQt5-sip==12.11.0
pyRFC3339==1.1
pyrsistent==0.15.5
pytest==4.6.9
python-dateutil==2.8.2
python-utils==3.1.0
pytz==2021.3
PyWavelets==1.2.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
rasterio==1.2.10
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.8
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.8.0
Shapely==1.8.0
simplejson==3.16.0
six==1.15.0
snuggs==1.4.7
soupsieve==1.9.5
stack-data==0.1.4
tensorboard==2.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.15.0
testpath==0.6.0
testresources==2.0.1
threadpoolctl==3.1.0
tifffile==2022.5.4
tomli==2.0.1
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.8
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wfdb==3.4.1
widgetsnbextension==3.5.2
win-unicode-console==0.5
wrapt==1.12.1
zipp==3.7.0
